Moved user methods into user controller
